// src/data/types/booking.ts
export interface Booking {
  id: string;
  userId: string;
  doctorId: string;
  doctorName: string;
  hospitalId: string;
  hospitalName: string;
  specialty: string;
  date: string;
  time: string;
  status: string;
  fee: number;
  patientName: string;
  dateOfBirth: string;
  gender: string;
  phone: string;
  email: string;
  zalo: string;
  address: string;
  reason: string;
  paymentMethod: string;
  bookingDate: string;
}